--- src/constants/index.js
@@ -1,3 +1,0 @@
-import types from './types.js';
-
-export { types };
