--- src/constants/types.js
@@ -1,12 +1,0 @@
-export default [
-  `gl`,
-  `lit`,
-  `m`,
-  `n`,
-  `phon`,
-  `sp`,
-  `tln`,
-  `trs`,
-  `txn`,
-  `w`,
-];
