+++ src/parseUtterances.js
@@ -2,6 +2,11 @@ import { getSchema }  from './utilities/index.js';
 import parseUtterance from './parseUtterance/index.js';
 
 /**
+ * A regular expression to match one or more empty lines
+ */
+const blankLinesRegExp = /(?:[ \t\v\u00a0\u1680\u2000-\u200a\u2028\u2029\u202f\u205f\u3000\ufeff]*\r?\n){2,}/gsu;
+
+/**
  * Finds the text of the utterances portion of a scription text and returns it
  * @param  {String} text The scription text
  * @return {String}
@@ -26,7 +31,6 @@ export default function parseUtterances(scription, codes, options) {
 
   if (!utterancesString) return [];
 
-  const blankLinesRegExp  = /(?:[ \t\v\u00a0\u1680\u2000-\u200a\u2028\u2029\u202f\u205f\u3000\ufeff]*\r?\n){2,}/gsu;
   const utterancesStrings = utterancesString.split(blankLinesRegExp);
   const schema            = getSchema(utterancesStrings[0]);
   const parse             = utteranceString => parseUtterance(utteranceString, schema, codes, options);
