+++ src/utilities/constants/brackets.js
@@ -1,0 +1,10 @@
+/**
+ * A hash of which brackets are used for each type of line
+ */
+export default {
+  lit:         [`[`, `]`],
+  phon:        [`[`, `]`],
+  tlnCurly:    [`‘`, `’`],
+  tlnStraight: [`'`, `'`],
+  txn:         [`/`, `/`],
+};
