+++ src/utilities/constants/lineCodes.js
@@ -1,0 +1,21 @@
+/* eslint-disable
+  sort-keys,
+*/
+
+/**
+ * The default set of backslash codes, in order. Do not adjust this ordering.
+ */
+export default {
+  sp:   `sp`,
+  trs:  `trs`,
+  txn:  `txn`,
+  w:    `w`,
+  phon: `phon`,
+  m:    `m`,
+  gl:   `gl`,
+  wlt:  `wlt`,
+  lit:  `lit`,
+  tln:  `tln`,
+  s:    `s`,
+  n:    `n`,
+};
