+++ src/utilities/constants/wordTypes.js
@@ -1,0 +1,4 @@
+/**
+ * An array of line types that are specific to words
+ */
+export default [`gl`, `m`, `w`, `wlt`];
