+++ src/utilities/index.js
@@ -1,31 +1,23 @@
-import difference          from './difference.js';
 import getCode             from './getCode.js';
 import getLines            from './getLines.js';
 import getLineType         from './getLineType.js';
 import getMatches          from './getMatches.js';
 import getSchema           from './getSchema.js';
 import groupLines          from './groupLines.js';
-import isString            from './isString.js';
-import isValidTag          from './isValidTag.js';
 import mergeTranscriptions from './mergeTranscriptions.js';
 import removeBrackets      from './removeBrackets.js';
 import removeEmphasis      from './removeEmphasis.js';
 import validateNumItems    from './validateNumItems.js';
-import zip                 from './zip.js';
 
 export {
-  difference,
   getCode,
   getLines,
   getLineType,
   getMatches,
   getSchema,
   groupLines,
-  isString,
-  isValidTag,
   mergeTranscriptions,
   removeBrackets,
   removeEmphasis,
   validateNumItems,
-  zip,
 };
