+++ src/utilities/js/difference.js
