+++ src/utilities/js/index.js
@@ -1,0 +1,2 @@
+export { default as difference } from './difference.js';
+export { default as zip } from './zip.js';
