+++ src/utilities/regexp/index.js
@@ -1,0 +1,1 @@
+export { default as newlineRegExp } from './newline.js';
