+++ src/utilities/regexp/newline.js
@@ -1,0 +1,4 @@
+/**
+ * Regular expression to match newlines
+ */
+export default /\r?\n/gu;
