+++ src/utilities/types/index.js
@@ -1,0 +1,3 @@
+export { default as isCode } from './isCode.js';
+export { default as isLanguageTag } from './isLanguageTag.js';
+export { default as isString } from './isString.js';
