+++ src/utilities/types/isLanguageTag.js
@@ -1,0 +1,11 @@
+/* eslint-disable
+  max-len,
+  prefer-named-capture-group,
+*/
+
+/**
+ * Checks whether a string is a valid IETF language tag. This RegExp was produced using the [ietf-language-tag-regex](https://github.com/sebinsua/ietf-language-tag-regex) library
+ */
+export default function validateLanguageTag(tag) {
+  return /^(((en-GB-oed|i-ami|i-bnn|i-default|i-enochian|i-hak|i-klingon|i-lux|i-mingo|i-navajo|i-pwn|i-tao|i-tay|i-tsu|sgn-BE-FR|sgn-BE-NL|sgn-CH-DE)|(art-lojban|cel-gaulish|no-bok|no-nyn|zh-guoyu|zh-hakka|zh-min|zh-min-nan|zh-xiang))|((([A-Za-z]{2,3}(-([A-Za-z]{3}(-[A-Za-z]{3}){0,2}))?)|[A-Za-z]{4}|[A-Za-z]{5,8})(-([A-Za-z]{4}))?(-([A-Za-z]{2}|[0-9]{3}))?(-([A-Za-z0-9]{5,8}|[0-9][A-Za-z0-9]{3}))*(-([0-9A-WY-Za-wy-z](-[A-Za-z0-9]{2,8})+))*(-(x(-[A-Za-z0-9]{1,8})+))?)|(x(-[A-Za-z0-9]{1,8})+))$/u.test(tag);
+}
