+++ src/utilities/types/isString.js
